pyproject.toml
[project]
name = "faster-whisper-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"fastapi>=0.110.0",
	"uvicorn[standard]>=0.29.0",
	"faster-whisper>=1.0.0",
	"python-multipart>=0.0.9",
	"pyyaml>=6.0.1",
]

[project.scripts]
faster-whisper-server = "faster_whisper_server.cli:main"
